+++ NEWS.md
@@ -1,4 +1,26 @@
-## v6.0.0 YYYY-mm-DD
+## vX.Y.Z YYYY-mm-DD
+### Breaking changes
+* Description ([ISSUE](https://folio-org.atlassian.net/browse/ISSUE))
+
+### New APIs versions
+* Provides `API_NAME vX.Y`
+* Requires `API_NAME vX.Y`
+
+### Features
+* Description ([ISSUE](https://folio-org.atlassian.net/browse/ISSUE))
+
+### Bug fixes
+* Description ([ISSUE](https://folio-org.atlassian.net/browse/ISSUE))
+
+### Tech Dept
+* Description ([ISSUE](https://folio-org.atlassian.net/browse/ISSUE))
+
+### Dependencies
+* Bump `LIB_NAME` from `OLD_VERSION` to `NEW_VERSION`
+* Add `LIB_NAME VERSION`
+* Remove `LIB_NAME`
+
+## v6.0.0 2026-04-14
 ### Breaking changes
 * Upgrade the module for Vert.X 5.0.x ([FST-94](https://issues.folio.org/browse/FST-94))
 * Upgrade the module for Spring Boot 4.0.x ([FST-102](https://issues.folio.org/browse/FST-102))